First Commit!
The code "save_cropped_raster.py" croppes the raster 'lidar_chm.tif' with the shapes "forma.shp" and saves in a compatible way with pytorch datasets. 
The utils module has a class to instantiate the pytorch dataset. 
the train.py trains a simple CNN example with the cropped raster, labeled with the shapefile atribute table.